Adds support for OAuth2 cherry-picked from the original repo.
This adds support for authentication using OAuth2.  This code
was cherry-picked from a pull request that was submitted to
the origina project's repo (not our own forked version.) That
pull is here:

  * https://github.com/nu7hatch/gmail/pull/80

This pull adds tests for OAuth2 that will work if credentials
for a valid account are entered.  Some of the tests will fail.
These are the same tests that fail for the original OAuth.